requirements.txt
fire==0.1.2
google-cloud-vision==0.27.0
Pillow==4.3.0
